task 6

+++ 0x05-pointers_arrays_strings/6-puts2.c
@@ -6,7 +6,7 @@
  */
 void puts2(char *str)
 {
-long int i;
+unsigned int i;
 
 
 for (i = 0; str[i] != '\0'; i++)
